python/requirements.txt
attrs==21.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-plugins==1.1.1
cligj==0.7.1
CouchDB==1.2
Fiona==1.8.19
geopandas==0.9.0
idna==2.10
joblib==1.0.1
munch==2.5.0
nltk==3.6.2
numpy==1.20.3
oauthlib==3.1.0
pandas==1.2.4
pkg-resources==0.0.0
pyproj==3.0.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
Shapely==1.7.1
six==1.16.0
tqdm==4.60.0
tweepy==3.10.0
urllib3==1.26.4
vaderSentiment==3.3.2
